fix: align orch-6 policy gates with the review-state model after base merge
The review-amount SUM gains the #524 ::float8 cast, the #440
enforcement scan documents main's ip-keyed login resend (#743) as a
direct-call exception (test allowlist + ADR-029 table), drops
/reminders from the must-use-helper list now that its rate-limited
bulk action no longer exists, and DesktopDashboard re-imports locale
for the #535 formatter signature.

+++ src/routes/(app)/dashboard/+page.server.ts
@@ -117,7 +117,7 @@ export const load: PageServerLoad = async ({ url, locals }) => {
 			trend, invoiceRangeRow,
 		] = await Promise.all([
 			db.select({
-				amount: sql<number>`COALESCE(SUM(CASE WHEN ${invoices.reviewState} <> 'revisado' THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)`,
+				amount: sql<number>`COALESCE(SUM(CASE WHEN ${invoices.reviewState} <> 'revisado' THEN COALESCE(${invoices.totalAmount},0) ELSE 0 END),0)::float8`,
 				count: sql<number>`COUNT(CASE WHEN ${invoices.reviewState} <> 'revisado' THEN 1 END)`,
 				incidencias: sql<number>`COUNT(CASE WHEN ${invoices.reviewState} = 'incidencia' THEN 1 END)`,
 			})
